LFN auto completion for rm.

+++ DataManagementSystem/Client/FileCatalogClientCLI.py
@@ -476,7 +476,42 @@ File Catalog Client $Revision: 1.17 $Date:
     """  
     # Not yet really implemented
     argss = args.split()
+    if len(argss) != 1:
+      print self.do_rm.__doc__
+      return
     self.removeFile(argss)
+
+  def complete_rmreplica(self, text, line, begidx, endidx):
+    result = []
+    args = line.split()
+
+    # the first argument -- LFN.
+    if (1<=len(args)<=2):
+      # If last char is ' ',
+      # this can be a new parameter.
+      if (len(args) == 1) or (len(args)==2 and (not line.endswith(' '))):
+        cur_path = ""
+        if (len(args) == 2):
+          cur_path = args[1]
+        result = self.lfn_dc.parse_text_line(text, cur_path, self.cwd)
+
+    return result
+
+  def complete_rm(self, text, line, begidx, endidx):
+    result = []
+    args = line.split()
+
+    # the first argument -- LFN.
+    if (1<=len(args)<=2):
+      # If last char is ' ',
+      # this can be a new parameter.
+      if (len(args) == 1) or (len(args)==2 and (not line.endswith(' '))):
+        cur_path = ""
+        if (len(args) == 2):
+          cur_path = args[1]
+        result = self.lfn_dc.parse_text_line(text, cur_path, self.cwd)
+
+    return result
     
   def do_rmdir(self,args):
     """ Remove directory from the storage and from the File Catalog
